LPS-65448 Move display logic to DisplayContext
