+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/resources/META-INF/resources/display/view.jsp
@@ -19,11 +19,11 @@
 <%
 String redirect = ParamUtil.getString(request, "redirect", currentURL);
 
-DDLRecordSet recordSet = ddlFormDisplayContext.getRecordSet();
+long recordSetId = ddlFormDisplayContext.getRecordSetId();
 %>
 
 <c:choose>
-	<c:when test="<%= (recordSet == null) %>">
+	<c:when test="<%= (recordSetId == 0) %>">
 		<div class="alert alert-info">
 			<liferay-ui:message key="select-an-existing-form-or-add-a-form-to-be-displayed-in-this-application" />
 		</div>
@@ -43,6 +43,10 @@ DDLRecordSet recordSet = ddlFormDisplayContext.getRecordSet();
 						<c:if test="<%= Validator.isNull(redirectURL) %>">
 							<aui:input name="redirect" type="hidden" value="<%= redirect %>" />
 						</c:if>
+
+						<%
+						DDLRecordSet recordSet = ddlFormDisplayContext.getRecordSet();
+						%>
 
 						<aui:input name="groupId" type="hidden" value="<%= recordSet.getGroupId() %>" />
 						<aui:input name="recordSetId" type="hidden" value="<%= recordSet.getRecordSetId() %>" />
@@ -109,7 +113,7 @@ DDLRecordSet recordSet = ddlFormDisplayContext.getRecordSet();
 						</div>
 
 						<div class="container-fluid-1280 ddl-form-builder-app">
-							<%= request.getAttribute(DDMWebKeys.DYNAMIC_DATA_MAPPING_FORM_HTML) %>
+							<%= ddlFormDisplayContext.getDDMFormHTML() %>
 						</div>
 					</aui:form>
 				</div>
